Polish debug panel: polling, split components, production guard
- Poll player states and full game state only when dialog is open
- Gate /api/debug/game/[gameId] behind NEXT_PUBLIC_DEBUG_MODE server-side
  (returns 404 when unset to avoid revealing endpoint existence)
- Split JsonTreeNode into its own file per one-component-per-file rule
- Remove unused defaultExpanded prop from JsonTreeNode
- Update CLAUDE.md: each component file must contain exactly one component
